+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Implement protocol on `Ipv4Addr` and `Ipv6Addr`, `(T0, )`, `()`, `Box`
 - Rename `Enum` trait to `EnumExt`
 - Delete `Settings`, replace with `ByteOrder`
+- Clean up `Error`
 # v0.2.0
 - Add context to all parse functions
 - Remove `#[repr(...)]`, instead specify repr in `#[protocol(discriminant = "...")]`